rename the auth provider types
