+++ server/test/auth/oidc/OidcCiviFormProfileAdapterTest.java
@@ -3,6 +3,7 @@ package auth.oidc;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import auth.ProfileFactory;
+import auth.oidc.applicant.IdcsProfileAdapter;
 import com.google.common.collect.ImmutableList;
 import java.util.Optional;
 import javax.inject.Provider;
@@ -20,7 +21,7 @@ public class OidcCiviFormProfileAdapterTest extends ResetPostgres {
   private static final String SUBJECT = "subject";
   private static final String AUTHORITY_ID = "iss: issuer sub: subject";
 
-  private OidcCiviFormProfileAdapter oidcProfileAdapter;
+  private OidcProfileAdapter oidcProfileAdapter;
   private ProfileFactory profileFactory;
 
   @Before
